pyproject.toml
[project]
name = "quickpod"
version = "0.1.0"
description = "Minimal GPU cluster reconciler (RunPod today; more vendors planned)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "runpod>=1.6.0",
  "PyYAML>=6.0",
  "pydantic>=2.0",
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "sqlalchemy>=2.0",
  "httpx>=0.27.0",
]

[project.scripts]
quickpod = "quickpod.__main__:main"

[project.optional-dependencies]
postgres = ["psycopg[binary]>=3.2"]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
